requirements.txt
discord.py
python-dotenv
gunicorn
flask